Extract performance contribution payload mapping
